Implement UI for existing functionality.

+++ src/main/java/com/github/amznlabs/swa_sample_seller/models/SellerAccessTokenResponseJson.java
@@ -4,5 +4,5 @@ public class SellerAccessTokenResponseJson {
     public String access_token;
     public String scope;
     public String token_type;
-    public int expires_in;
+    public Long expires_in;
 }
