+++ src/main/java/com/github/amznlabs/swa_sample_seller/models/SubscriptionsResponseJson.java
@@ -3,10 +3,10 @@ package com.github.amznlabs.swa_sample_seller.models;
 import java.util.List;
 
 public class SubscriptionsResponseJson {
-    public Integer nextCursor;
+    public Long nextCursor;
     public List<SubscriptionJson> subscriptions;
     public String access_token;
     public String scope;
     public String token_type;
-    public int expires_in;
+    public Long expires_in;
 }
